requirements.txt
filestack-python>=1.0.0
requests>=2.28.0
tkinter-tooltip>=2.0.0